feat: move css engine loader to shared folder
