+++ apps/designer/app/shared/css-utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./delete-css-rules-by-breakpoint";
+export * from "./generate-css-text";
